requirements.txt
Flask==3.1.0
requests==2.32.3
PyYAML==6.0.2
redis==5.0.7
